+++ source/renderer/app/containers/wallet/WalletReceivePage.js
@@ -1,6 +1,7 @@
 // @flow
 import React, { Component, Fragment } from 'react';
 import { observer, inject } from 'mobx-react';
+import { showSaveDialogChannel } from '../../ipc/show-file-dialog-channels';
 import WalletReceiveRandom from '../../components/wallet/receive/WalletReceiveRandom';
 import WalletReceiveSequential from '../../components/wallet/receive/WalletReceiveSequential';
 import VerticalFlexContainer from '../../components/layout/VerticalFlexContainer';
@@ -56,7 +57,7 @@ export default class WalletReceivePage extends Component<Props, State> {
     this.props.actions.dialogs.closeActiveDialog.trigger();
   };
 
-  handleDownloadPDF = (note: string) => {
+  handleDownloadPDF = async (note: string) => {
     const { addressToShare } = this.state;
 
     const name = generateFileNameWithTimestamp({
@@ -65,8 +66,7 @@ export default class WalletReceivePage extends Component<Props, State> {
       isUTC: false,
     });
 
-    // TODO: refactor this direct access to the dialog api
-    const filePath = global.dialog.showSaveDialog({
+    const params = {
       defaultPath: `${name}.pdf`,
       filters: [
         {
@@ -74,7 +74,8 @@ export default class WalletReceivePage extends Component<Props, State> {
           extensions: ['pdf'],
         },
       ],
-    });
+    };
+    const { filePath } = await showSaveDialogChannel.send(params);
 
     // if cancel button is clicked or path is empty
     if (!filePath || !addressToShare) return;