+++ source/renderer/app/utils/logging.js
@@ -3,11 +3,10 @@ import { formatContext } from '../../../common/utils/logging';
 import type {
   FormatMessageContextParams,
   Logger,
+  LoggingLevel,
 } from '../../../common/types/logging.types';
 
-const log = global.electronLog;
-const { environment } = global;
-
+const { environment, electronLog } = global;
 const appName = 'daedalus';
 const electronProcess = 'ipcRenderer';
 const { network, os, platformVersion, version } = environment;
@@ -26,12 +25,20 @@ const environmentData = {
   version,
 };
 
-const logToLevel = (level: string) => (message: string, data: ?Object) =>
-  log[level](formatContext({ ...messageContext, level }), {
-    message,
-    data,
-    environmentData,
-  });
+const logToLevel = (level: LoggingLevel) => (
+  message: string,
+  data: ?Object
+) => {
+  const args = [
+    formatContext({ ...messageContext, level }),
+    {
+      message,
+      data,
+      environmentData,
+    },
+  ];
+  electronLog[level](...args);
+};
 
 export const logger: Logger = {
   debug: logToLevel('debug'),